+++ crates/config/src/patterns.rs
@@ -22,11 +22,11 @@ pattern!(TOKEN_FUNC, "@([a-z]+)\\(([0-9A-Za-z_-]+)\\)");
 pattern!(TOKEN_FUNC_DISTINCT, "^@([a-z]+)\\(([0-9A-Za-z_-]+)\\)$");
 pattern!(
     TOKEN_VAR,
-    "\\$(arch|language|osFamily|os|projectAlias|projectChannel|projectName|projectLayer|projectOwner|projectRoot|projectSource|projectStack|projectType|project|target|taskPlatform|taskToolchain|taskToolchains|taskType|task|timestamp|datetime|date|time|vcsBranch|vcsRepository|vcsRevision|workingDir|workspaceRoot)"
+    "\\$(arch|language|osFamily|os|projectAlias|projectChannel|projectName|projectLayer|projectOwner|projectRoot|projectSource|projectStack|projectType|project|target|taskToolchain|taskToolchains|taskType|task|timestamp|datetime|date|time|vcsBranch|vcsRepository|vcsRevision|workingDir|workspaceRoot)"
 );
 pattern!(
     TOKEN_VAR_DISTINCT,
-    "^\\$(arch|language|osFamily|os|projectAlias|projectChannel|projectName|projectLayer|projectOwner|projectRoot|projectSource|projectStack|projectType|project|target|taskPlatform|taskToolchain|taskToolchains|taskType|task|timestamp|datetime|date|time|vcsBranch|vcsRepository|vcsRevision|workingDir|workspaceRoot)$"
+    "^\\$(arch|language|osFamily|os|projectAlias|projectChannel|projectName|projectLayer|projectOwner|projectRoot|projectSource|projectStack|projectType|project|target|taskToolchain|taskToolchains|taskType|task|timestamp|datetime|date|time|vcsBranch|vcsRepository|vcsRevision|workingDir|workspaceRoot)$"
 );
 
 pub fn merge_iter<I, V, C>(mut prev: I, next: I, _: &C) -> MergeResult<I>