+++ crates/query/src/builder.rs
@@ -1,12 +1,11 @@
 use crate::parser::{AstNode, ComparisonOperator, LogicalOperator, parse_query};
 use crate::query_error::QueryError;
-use moon_common::color;
 use moon_config::{LanguageType, LayerType, StackType, TaskType};
 use starbase_utils::glob::GlobSet;
 use std::borrow::Cow;
 use std::cmp::PartialEq;
 use std::str::FromStr;
-use tracing::{debug, instrument};
+use tracing::instrument;
 
 pub type FieldValue<'l> = Cow<'l, str>;
 pub type FieldValues<'l> = Vec<FieldValue<'l>>;
@@ -23,7 +22,6 @@ pub enum Field<'l> {
     ProjectType(Vec<LayerType>),
     Tag(FieldValues<'l>),
     Task(FieldValues<'l>),
-    TaskPlatform(FieldValues<'l>),
     TaskToolchain(FieldValues<'l>),
     TaskType(Vec<TaskType>),
 }
@@ -136,15 +134,6 @@ fn build_criteria(ast: Vec<AstNode<'_>>) -> miette::Result<Criteria<'_>> {
                     }
                     "tag" => Field::Tag(value),
                     "task" => Field::Task(value),
-                    "taskPlatform" => {
-                        debug!(
-                            "The {} query field is deprecated, use {} instead",
-                            color::property("taskPlatform"),
-                            color::property("taskToolchain"),
-                        );
-
-                        Field::TaskPlatform(value)
-                    }
                     "taskToolchain" => Field::TaskToolchain(value),
                     "taskType" => {
                         Field::TaskType(build_criteria_enum::<TaskType>(&field, &op, value)?)