+++ crates/workspace-graph/src/query_tasks.rs
@@ -85,7 +85,7 @@ fn does_task_match_criteria(task: &Task, query: &Criteria) -> miette::Result<boo
                         }
                     }
                     Field::Task(ids) => condition.matches(ids, &task.id),
-                    Field::TaskPlatform(ids) | Field::TaskToolchain(ids) => {
+                    Field::TaskToolchain(ids) => {
                         let toolchains = task
                             .toolchains
                             .iter()